requirements.txt
ultralytics==8.1.0
torch==2.6.0
torchvision==0.21.0
pillow
numpy
ensemble-boxes>=1.0.9
